third_party/blink/web_tests/external/wpt/css/css-color/parsing/color-valid-color-mix-function-expected.txt

This is a testharness.js-based test.
Found 43 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] e.style['color'] = "color-mix(hsl(120deg 10% 20%), hsl(30deg 30% 40%))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color-mix(hwb(120deg 10% 20%), hwb(30deg 30% 40%))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color-mix(lch(10 20 30), lch(50 60 70))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color-mix(oklch(0.1 20 30), oklch(0.5 60 70))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color-mix(lab(10 20 30), lab(50 60 70))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(0.1 0.2 0.3), oklab(0.5 0.6 0.7))" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3), oklab(0.5 0.6 0.7))\nExpected: color-mix(oklab(0.1 0.2 0.3), oklab(0.5 0.6 0.7)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(  ), oklab(  ))" but got "color-mix(in oklab, oklab(  ), oklab(  ))"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(0.1 0.2 0.3) 25%, oklab(0.5 0.6 0.7))" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3) 25%, oklab(0.5 0.6 0.7))\nExpected: color-mix(oklab(0.1 0.2 0.3) 25%, oklab(0.5 0.6 0.7)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(  ) %, oklab(  ))" but got "color-mix(in oklab, oklab(  ) %, oklab(  ))"
[FAIL] e.style['color'] = "color-mix(in oklab, 25% oklab(0.1 0.2 0.3), oklab(0.5 0.6 0.7))" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3) 25%, oklab(0.5 0.6 0.7))\nExpected: color-mix(oklab(0.1 0.2 0.3) 25%, oklab(0.5 0.6 0.7)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(  ) %, oklab(  ))" but got "color-mix(in oklab, oklab(  ) %, oklab(  ))"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(0.1 0.2 0.3), 25% oklab(0.5 0.6 0.7))" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3) 75%, oklab(0.5 0.6 0.7))\nExpected: color-mix(oklab(0.1 0.2 0.3) 75%, oklab(0.5 0.6 0.7)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(  ) %, oklab(  ))" but got "color-mix(in oklab, oklab(  ) %, oklab(  ))"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(0.1 0.2 0.3), oklab(0.5 0.6 0.7) 25%)" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3) 75%, oklab(0.5 0.6 0.7))\nExpected: color-mix(oklab(0.1 0.2 0.3) 75%, oklab(0.5 0.6 0.7)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(  ) %, oklab(  ))" but got "color-mix(in oklab, oklab(  ) %, oklab(  ))"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(0.1 0.2 0.3) 25%, oklab(0.5 0.6 0.7) 75%)" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3) 25%, oklab(0.5 0.6 0.7))\nExpected: color-mix(oklab(0.1 0.2 0.3) 25%, oklab(0.5 0.6 0.7)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(  ) %, oklab(  ))" but got "color-mix(in oklab, oklab(  ) %, oklab(  ))"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(0.1 0.2 0.3) 30%, oklab(0.5 0.6 0.7) 90%)" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3) 30%, oklab(0.5 0.6 0.7) 90%)\nExpected: color-mix(oklab(0.1 0.2 0.3) 30%, oklab(0.5 0.6 0.7) 90%).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(  ) %, oklab(  ) %)" but got "color-mix(in oklab, oklab(  ) %, oklab(  ) %)"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(0.1 0.2 0.3) 12.5%, oklab(0.5 0.6 0.7) 37.5%)" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3) 12.5%, oklab(0.5 0.6 0.7) 37.5%)\nExpected: color-mix(oklab(0.1 0.2 0.3) 12.5%, oklab(0.5 0.6 0.7) 37.5%).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(  ) %, oklab(  ) %)" but got "color-mix(in oklab, oklab(  ) %, oklab(  ) %)"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(0.1 0.2 0.3) 0%, oklab(0.5 0.6 0.7))" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3) 0%, oklab(0.5 0.6 0.7))\nExpected: color-mix(oklab(0.1 0.2 0.3) 0%, oklab(0.5 0.6 0.7)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(  ) %, oklab(  ))" but got "color-mix(in oklab, oklab(  ) %, oklab(  ))"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(0.1 0.2 0.3 / .4), oklab(0.5 0.6 0.7 / .8))" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3 / 0.4), oklab(0.5 0.6 0.7 / 0.8))\nExpected: color-mix(oklab(0.1 0.2 0.3 / 0.4), oklab(0.5 0.6 0.7 / 0.8)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(   / ), oklab(   / ))" but got "color-mix(in oklab, oklab(   / ), oklab(   / ))"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(0.1 0.2 0.3 / .4) 25%, oklab(0.5 0.6 0.7 / .8))" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3 / 0.4) 25%, oklab(0.5 0.6 0.7 / 0.8))\nExpected: color-mix(oklab(0.1 0.2 0.3 / 0.4) 25%, oklab(0.5 0.6 0.7 / 0.8)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(   / ) %, oklab(   / ))" but got "color-mix(in oklab, oklab(   / ) %, oklab(   / ))"
[FAIL] e.style['color'] = "color-mix(in oklab, 25% oklab(0.1 0.2 0.3 / .4), oklab(0.5 0.6 0.7 / .8))" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3 / 0.4) 25%, oklab(0.5 0.6 0.7 / 0.8))\nExpected: color-mix(oklab(0.1 0.2 0.3 / 0.4) 25%, oklab(0.5 0.6 0.7 / 0.8)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(   / ) %, oklab(   / ))" but got "color-mix(in oklab, oklab(   / ) %, oklab(   / ))"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(0.1 0.2 0.3 / .4), 25% oklab(0.5 0.6 0.7 / .8))" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3 / 0.4) 75%, oklab(0.5 0.6 0.7 / 0.8))\nExpected: color-mix(oklab(0.1 0.2 0.3 / 0.4) 75%, oklab(0.5 0.6 0.7 / 0.8)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(   / ) %, oklab(   / ))" but got "color-mix(in oklab, oklab(   / ) %, oklab(   / ))"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(0.1 0.2 0.3 / .4), oklab(0.5 0.6 0.7 / .8) 25%)" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3 / 0.4) 75%, oklab(0.5 0.6 0.7 / 0.8))\nExpected: color-mix(oklab(0.1 0.2 0.3 / 0.4) 75%, oklab(0.5 0.6 0.7 / 0.8)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(   / ) %, oklab(   / ))" but got "color-mix(in oklab, oklab(   / ) %, oklab(   / ))"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(0.1 0.2 0.3 / .4) 25%, oklab(0.5 0.6 0.7 / .8) 75%)" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3 / 0.4) 25%, oklab(0.5 0.6 0.7 / 0.8))\nExpected: color-mix(oklab(0.1 0.2 0.3 / 0.4) 25%, oklab(0.5 0.6 0.7 / 0.8)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(   / ) %, oklab(   / ))" but got "color-mix(in oklab, oklab(   / ) %, oklab(   / ))"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(0.1 0.2 0.3 / .4) 30%, oklab(0.5 0.6 0.7 / .8) 90%)" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3 / 0.4) 30%, oklab(0.5 0.6 0.7 / 0.8) 90%)\nExpected: color-mix(oklab(0.1 0.2 0.3 / 0.4) 30%, oklab(0.5 0.6 0.7 / 0.8) 90%).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(   / ) %, oklab(   / ) %)" but got "color-mix(in oklab, oklab(   / ) %, oklab(   / ) %)"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(0.1 0.2 0.3 / .4) 12.5%, oklab(0.5 0.6 0.7 / .8) 37.5%)" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3 / 0.4) 12.5%, oklab(0.5 0.6 0.7 / 0.8) 37.5%)\nExpected: color-mix(oklab(0.1 0.2 0.3 / 0.4) 12.5%, oklab(0.5 0.6 0.7 / 0.8) 37.5%).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(   / ) %, oklab(   / ) %)" but got "color-mix(in oklab, oklab(   / ) %, oklab(   / ) %)"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(0.1 0.2 0.3 / .4) 0%, oklab(0.5 0.6 0.7 / .8))" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3 / 0.4) 0%, oklab(0.5 0.6 0.7 / 0.8))\nExpected: color-mix(oklab(0.1 0.2 0.3 / 0.4) 0%, oklab(0.5 0.6 0.7 / 0.8)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(   / ) %, oklab(   / ))" but got "color-mix(in oklab, oklab(   / ) %, oklab(   / ))"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(none none none), oklab(none none none))" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(none none none), oklab(none none none))\nExpected: color-mix(oklab(none none none), oklab(none none none)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(none none none), oklab(none none none))" but got "color-mix(in oklab, oklab(none none none), oklab(none none none))"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(none none none), oklab(0.5 0.6 0.7))" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(none none none), oklab(0.5 0.6 0.7))\nExpected: color-mix(oklab(none none none), oklab(0.5 0.6 0.7)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(none none none), oklab(  ))" but got "color-mix(in oklab, oklab(none none none), oklab(  ))"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(0.1 0.2 0.3), oklab(none none none))" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3), oklab(none none none))\nExpected: color-mix(oklab(0.1 0.2 0.3), oklab(none none none)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(  ), oklab(none none none))" but got "color-mix(in oklab, oklab(  ), oklab(none none none))"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(0.1 0.2 none), oklab(0.5 0.6 0.7))" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 none), oklab(0.5 0.6 0.7))\nExpected: color-mix(oklab(0.1 0.2 none), oklab(0.5 0.6 0.7)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(  none), oklab(  ))" but got "color-mix(in oklab, oklab(  none), oklab(  ))"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(0.1 0.2 0.3), oklab(0.5 0.6 none))" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3), oklab(0.5 0.6 none))\nExpected: color-mix(oklab(0.1 0.2 0.3), oklab(0.5 0.6 none)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(  ), oklab(  none))" but got "color-mix(in oklab, oklab(  ), oklab(  none))"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(none 0.2 0.3), oklab(0.5 none 0.7))" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(none 0.2 0.3), oklab(0.5 none 0.7))\nExpected: color-mix(oklab(none 0.2 0.3), oklab(0.5 none 0.7)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(none  ), oklab( none ))" but got "color-mix(in oklab, oklab(none  ), oklab( none ))"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(0.1 0.2 0.3 / none), oklab(0.5 0.6 0.7))" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3 / none), oklab(0.5 0.6 0.7))\nExpected: color-mix(oklab(0.1 0.2 0.3 / none), oklab(0.5 0.6 0.7)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(   / none), oklab(  ))" but got "color-mix(in oklab, oklab(   / none), oklab(  ))"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(0.1 0.2 0.3 / none), oklab(0.5 0.6 0.7 / 0.5))" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3 / none), oklab(0.5 0.6 0.7 / 0.5))\nExpected: color-mix(oklab(0.1 0.2 0.3 / none), oklab(0.5 0.6 0.7 / 0.5)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(   / none), oklab(   / ))" but got "color-mix(in oklab, oklab(   / none), oklab(   / ))"
[FAIL] e.style['color'] = "color-mix(in oklab, oklab(0.1 0.2 0.3 / none), oklab(0.5 0.6 0.7 / none))" should set the property value
  Colors do not match.\nActual:   color-mix(in oklab, oklab(0.1 0.2 0.3 / none), oklab(0.5 0.6 0.7 / none))\nExpected: color-mix(oklab(0.1 0.2 0.3 / none), oklab(0.5 0.6 0.7 / none)).\nError: assert_equals: Color format is correct. expected "color-mix(oklab(   / none), oklab(   / none))" but got "color-mix(in oklab, oklab(   / none), oklab(   / none))"
[FAIL] e.style['color'] = "color-mix(oklab(0.1 0.2 0.3), oklab(0.5 0.6 0.7))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color-mix(color(srgb .1 .2 .3), color(srgb .5 .6 .7))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color-mix(color(srgb-linear .1 .2 .3), color(srgb-linear .5 .6 .7))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color-mix(color(display-p3 .1 .2 .3), color(display-p3 .5 .6 .7))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color-mix(color(display-p3-linear .1 .2 .3), color(display-p3-linear .5 .6 .7))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color-mix(color(a98-rgb .1 .2 .3), color(a98-rgb .5 .6 .7))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color-mix(color(prophoto-rgb .1 .2 .3), color(prophoto-rgb .5 .6 .7))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color-mix(color(rec2020 .1 .2 .3), color(rec2020 .5 .6 .7))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color-mix(color(xyz .1 .2 .3), color(xyz .5 .6 .7))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color-mix(color(xyz-d50 .1 .2 .3), color(xyz-d50 .5 .6 .7))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['color'] = "color-mix(color(xyz-d65 .1 .2 .3), color(xyz-d65 .5 .6 .7))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
